Refactor into OOP and implement Thor
I don't think the CURRENT requirements would _necessitate_ the new design, but I think these changes should make it easier to expand the requirements in the future when the need for change arises. I am certainly open to improvements on this design!

This change makes use of a more robust command line solution (thor) and
makes the solution more maintainable by introducing an oop design. Also
will require a rewrite on some of the documentation (which this commit
does not include.)

* The Thorfile requires Thor to run. (gem install thor). It provides two
tasks, `Thor page:new TITLE CATEGORY` and `Thor page:relink`. To see all
the tasks and their descriptions, run `Thor list` in the terminal.
* A few notes on the page:new task- it now requires both title and
* category. Before they were optional. Also, Thor page:relink deletes all the files in _pages and regenerates them.
* The lib/helpers.rb file provides a Helpers module and a Page class. This
module could be expanded to include other helper classes.